requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

pillow==10.3.0
tokenizers==0.19.1
torch==2.3.0
pytest==8.2.1
numpy==1.26.4
scikit_learn==1.5.0
transformers==4.41.1
